refactor: split distribution into 3 phases for manual signing steps
Replace single distribute.ts with phase1/phase2/phase3 scripts that
pass state via a JSON file, removing the need for interactive stdin
pauses between Safe signing steps.

Also updates README with full distribution docs and .env.example
with all required variables.
