+++ .env.example
@@ -2,3 +2,8 @@ SNAPSHOT_BLOCK_1=
 SNAPSHOT_BLOCK_2=
 HYPERSYNC_API_TOKEN=
 PROPOSER_PRIVATE_KEY=
+BASE_RPC_URL=
+PINATA_JWT=
+SAFE_API_KEY=
+GITHUB_TOKEN=
+ISSUANCE_SITE_PATH=
